Build with msvc 2017 rc 1.

+++ source/Core/Base/include/BaseTypes.h
@@ -24,7 +24,7 @@ typedef std::uint8_t uint8;
 typedef std::uint16_t uint16;
 typedef std::uint32_t uint32;
 typedef std::uint64_t uint64;
-typedef std::size_t hash64;
+typedef std::uint64_t hash64;
 
 typedef uint8 byte;
 typedef uint8 utf8;
